fix some config

+++ Ighan.AutoDelivery.ConsoleDemo/Utilities/DirectoryFullGraphCopier.cs
@@ -11,13 +11,23 @@ namespace Ighan.AutoDelivery.Core
         {
             Console.WriteLine($"Start: coping => {sourcePath} into: {destinationPath}");
 
-            CreateDirectory(new DirectoryInfo(destinationPath));
+            if (Directory.Exists(sourcePath))
+            {
+                CreateDirectory(new DirectoryInfo(destinationPath));
 
-            foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
-                Directory.CreateDirectory(dirPath.Replace(sourcePath, destinationPath));
+                foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
+                    Directory.CreateDirectory(dirPath.Replace(sourcePath, destinationPath));
 
-            foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
-                File.Copy(newPath, newPath.Replace(sourcePath, destinationPath), true);
+                foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
+                    File.Copy(newPath, newPath.Replace(sourcePath, destinationPath), true);
+            }
+            else
+            {
+                var defaultColor = Console.ForegroundColor;
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("Directory(source) not exists: " + sourcePath);
+                Console.ForegroundColor = defaultColor;
+            }
 
             Console.WriteLine($"End: coping");
         }
